Extract test utils out of message_assembly and load metadata stub
