+++ circuit/src/execution-delivery/src/message_assembly/substrate_gateway_protocol.rs
@@ -401,122 +401,31 @@ where
 }
 
 #[cfg(test)]
-mod tests {
+pub mod tests {
     use codec::Encode;
-    use frame_metadata::{
-        DecodeDifferent, ExtrinsicMetadata, FunctionMetadata, ModuleMetadata, RuntimeMetadataV13,
-    };
+    
     use sp_application_crypto::RuntimePublic;
     use sp_core::sr25519::Signature;
-    use sp_core::H256;
+    
     use sp_io::TestExternalities;
     use sp_keystore::testing::KeyStore;
     use sp_keystore::{KeystoreExt, SyncCryptoStore};
-    use sp_version::{ApisVec, RuntimeVersion};
+    
 
     use t3rn_primitives::transfers::TransferEntry;
     use t3rn_primitives::GatewayType;
 
-    use crate::crypto::Public;
+    
     use crate::KEY_TYPE;
+
+    use crate::message_assembly::test_utils::*;
 
     use super::{
         CircuitOutboundMessage, GatewayExpectedOutput, GatewayInboundProtocol, MessagePayload,
-        Metadata, SubstrateGatewayProtocol,
     };
 
-    fn create_test_metadata(
-        modules_with_functions: Vec<(&'static str, Vec<&'static str>)>,
-    ) -> Metadata {
-        let mut module_index = 1;
-        let mut modules: Vec<ModuleMetadata> = vec![];
 
-        let fn_metadata_generator = |name: &'static str| -> FunctionMetadata {
-            FunctionMetadata {
-                name: DecodeDifferent::Encode(name),
-                arguments: DecodeDifferent::Decoded(vec![]),
-                documentation: DecodeDifferent::Decoded(vec![]),
-            }
-        };
-
-        let module_metadata_generator = |mod_name: &'static str,
-                                         mod_index: u8,
-                                         functions: Vec<FunctionMetadata>|
-         -> ModuleMetadata {
-            ModuleMetadata {
-                index: mod_index,
-                name: DecodeDifferent::Encode(mod_name),
-                storage: None,
-                calls: Some(DecodeDifferent::Decoded(functions)),
-                event: None,
-                constants: DecodeDifferent::Decoded(vec![]),
-                errors: DecodeDifferent::Decoded(vec![]),
-            }
-        };
-
-        for module in modules_with_functions {
-            let (module_name, fn_names) = module;
-            let functions = fn_names.into_iter().map(fn_metadata_generator).collect();
-            modules.push(module_metadata_generator(
-                module_name,
-                module_index,
-                functions,
-            ));
-            module_index = module_index + 1;
-        }
-
-        let runtime_metadata = RuntimeMetadataV13 {
-            extrinsic: ExtrinsicMetadata {
-                version: 1,
-                signed_extensions: vec![DecodeDifferent::Decoded(String::from("test"))],
-            },
-            modules: DecodeDifferent::Decoded(modules),
-        };
-        Metadata::new(runtime_metadata)
-    }
-
-    fn create_test_runtime_version() -> RuntimeVersion {
-        RuntimeVersion {
-            spec_name: "circuit-runtime".into(),
-            impl_name: "circuit-runtime".into(),
-            authoring_version: 1,
-            impl_version: 1,
-            apis: ApisVec::Owned(vec![([0_u8; 8], 0_u32)]),
-            transaction_version: 4,
-            spec_version: 13,
-        }
-    }
-
-    fn create_submitter() -> Public {
-        Public::default()
-    }
-
-    fn create_test_genesis_hash() -> H256 {
-        [0_u8; 32].into()
-    }
-
-    fn create_default_test_gateway_protocol() -> SubstrateGatewayProtocol<Public, H256> {
-        SubstrateGatewayProtocol::new(
-            Metadata::default(),
-            create_test_runtime_version(),
-            create_test_genesis_hash(),
-            create_submitter(),
-        )
-    }
-
-    fn create_test_gateway_protocol(
-        modules_with_functions: Vec<(&'static str, Vec<&'static str>)>,
-        submitter: Public,
-    ) -> SubstrateGatewayProtocol<Public, H256> {
-        SubstrateGatewayProtocol::new(
-            create_test_metadata(modules_with_functions),
-            create_test_runtime_version(),
-            create_test_genesis_hash(),
-            submitter,
-        )
-    }
-
-    fn assert_signed_payload(
+    pub fn assert_signed_payload(
         actual: CircuitOutboundMessage,
         submitter: sp_core::sr25519::Public,
         exp_arguments: Vec<Vec<u8>>,
@@ -807,6 +716,8 @@ mod tests {
 
         let mut ext = TestExternalities::new_empty();
         ext.register_extension(KeystoreExt(keystore.into()));
+
+
         ext.execute_with(|| {
             let test_protocol = create_test_gateway_protocol(
                 vec![("gatewayEscrowed", vec!["callStatic"])],
@@ -1002,6 +913,7 @@ mod tests {
 
         let mut ext = TestExternalities::new_empty();
         ext.register_extension(KeystoreExt(keystore.into()));
+
         ext.execute_with(|| {
             let test_protocol = create_test_gateway_protocol(
                 vec![
